requirements.txt
# 基础依赖
flask==2.2.3
requests==2.28.2
beautifulsoup4==4.12.2
python-dotenv==1.0.0

# 数据处理（可选）
pandas==1.5.3

# Web服务相关
gunicorn==20.1.0  # 用于生产环境部署